% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageFirstDiagnosisEstimation.R
\name{ageFirstDiagnosis}
\alias{ageFirstDiagnosis}
\title{An new function to visualize and analyze the age of the first diagnosis given a list of disorders}
\usage{
ageFirstDiagnosis(input, diseaseList, sex, cutOff = 5,
  visualization = FALSE, fisherTest = TRUE, fisherCutOff = 0.001)
}
\description{
An new function to visualize and analyze the age of the first diagnosis given a list of disorders
}
