% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beforeAfterASD.R
\name{beforeAfterASD}
\alias{beforeAfterASD}
\title{A new function to analyze if a phenotype is diagnosed before or after ASD}
\usage{
beforeAfterASD(input, autismCodes, phenotypeList)
}
\description{
A new function to analyze if a phenotype is diagnosed before or after ASD
}
