% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapDirectionAdapted.R
\name{heatmapDirectionAdapted}
\alias{heatmapDirectionAdapted}
\title{An adapted function of the heatmapDirection}
\usage{
heatmapDirectionAdapted(input, fromAtoBColor = "darkgreen",
  fromBtoAColor = "orange", noDirectionColor = "grey", verbose = FALSE,
  warnings = TRUE, interactive = FALSE)
}
\description{
An adapted function of the heatmapDirection
}
