% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapPlotAdaptedTriangle.R
\name{heatmapPlotAdaptedTriangle}
\alias{heatmapPlotAdaptedTriangle}
\title{An adapted function of the heatmapPlot}
\usage{
heatmapPlotAdaptedTriangle(input, selectValue = "score", cutOff = 1,
  npairs = 1, interactive = FALSE, lowColor = "#0000FF",
  highColor = "yellow", verbose = FALSE)
}
\description{
An adapted function of the heatmapPlot
}
