% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapSexRatioAdapted.R
\name{heatmapSexRatioAdapted}
\alias{heatmapSexRatioAdapted}
\title{An adapted function of the heatmapSexRatio}
\usage{
heatmapSexRatioAdapted(input, highColor = "blue", lowColor = "red",
  verbose = FALSE, warnings = TRUE, interactive = FALSE)
}
\description{
An adapted function of the heatmapSexRatio
}
