% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalenceEstimation.R
\name{prevalenceEstimation}
\alias{prevalenceEstimation}
\title{An new function to visualize and analyze the prevalence of disorders}
\usage{
prevalenceEstimation(input, sex, cutOff = 5, visualization = FALSE,
  fisherTest = TRUE, fisherCutOff = 0.001)
}
\description{
An new function to visualize and analyze the prevalence of disorders
}
